+++ tasks/social-data-anomaly-report/tests/verify.py
@@ -28,10 +28,23 @@ addressed email with an empty or off-topic body must not pass.
 import sqlite3
 import sys
 
-sys.path.insert(0, "/workspace/environment")
-from verify_utils import token_f1
-
 SOCIAL_DB = "/opt/mock/data/social/social.db"
+
+
+def token_f1(expected: str, actual: str) -> float:
+    """Compute token-level F1 between two strings."""
+    expected_tokens = set(expected.lower().split())
+    actual_tokens = set(actual.lower().split())
+    if not expected_tokens or not actual_tokens:
+        return 0.0
+    matched = expected_tokens & actual_tokens
+    precision = len(matched) / len(actual_tokens)
+    recall = len(matched) / len(expected_tokens)
+    if precision + recall == 0:
+        return 0.0
+    return 2 * precision * recall / (precision + recall)
+
+
 EMAIL_DB = "/var/lib/mock-data/email/email.db"
 
 EXPECTED_RECIPIENT = "data-team@mosi.inc"
