+++ tasks/social-event-campaign/tests/verify.py
@@ -24,10 +24,23 @@ import sys
 import urllib.error
 import urllib.request
 
-sys.path.insert(0, "/workspace/environment")
-from verify_utils import token_f1
-
 SOCIAL_URL = "http://127.0.0.1:5008"
+
+
+def token_f1(expected: str, actual: str) -> float:
+    """Compute token-level F1 between two strings."""
+    expected_tokens = set(expected.lower().split())
+    actual_tokens = set(actual.lower().split())
+    if not expected_tokens or not actual_tokens:
+        return 0.0
+    matched = expected_tokens & actual_tokens
+    precision = len(matched) / len(actual_tokens)
+    recall = len(matched) / len(expected_tokens)
+    if precision + recall == 0:
+        return 0.0
+    return 2 * precision * recall / (precision + recall)
+
+
 EMAIL_DB_PATH = "/var/lib/mock-data/email/email.db"
 
 SOCIAL_USERNAME = "mosi_brand"
